* ext/zlib/zlib.c (gzfile_check_footer): ISIZE (Input SIZE) in   gzip's header is the size of uncompressed input data modulo 2^32.   [ruby-core:34481] http://www.ietf.org/rfc/rfc1952.txt
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@30540 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/zlib/zlib.c
@@ -2169,7 +2169,7 @@ gzfile_check_footer(struct gzfile *gz)
     if (gz->crc != crc) {
 	rb_raise(cCRCError, "invalid compressed data -- crc error");
     }
-    if (gz->z.stream.total_out != length) {
+    if ((int32_t)gz->z.stream.total_out != length) {
 	rb_raise(cLengthError, "invalid compressed data -- length error");
     }
 }
